Complete the User package and create an api that will handle the getAllUsers

+++ src/main/java/com/store/discounts/user/UserRepository.java
@@ -1,9 +1,12 @@
 package com.store.discounts.user;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 }
